test(integration): split IntegrationTests into 3 focused classes (test-suite-integration-class-split)
Splits the 451-line `IntegrationTests` kitchen-sink class into three focused
classes with the same `[DoNotParallelize][TestClass]` + `SharedWorkspaceTestBase`
inheritance pattern:

- `IntegrationTests_WorkspaceCore` (8 tests) — workspace metadata, status,
  project graph serialization.
- `IntegrationTests_SymbolNavigation` (14 tests) — symbol search/info,
  go-to-definition, find-references, find-implementations, document symbols,
  type hierarchy, callers/callees, impact analysis, diagnostics.
- `IntegrationTests_EditApply` (6 tests) — format/organize-usings preview +
  apply on isolated workspace copies, rename preview/apply, preview-token
  staleness regression.

No assertions changed; all 28 `[TestMethod]` bodies moved verbatim. Shared
workspace cache (`GetOrLoadWorkspaceIdAsync(SampleSolutionPath, ...)`) means
each class's `[ClassInitialize]` reuses the existing workspace — no per-class
load cost. `ClassCleanup` calls `DisposeServices()` which is an intentional
no-op (assembly disposal lives in `AssemblyLifecycle.Cleanup`), so three
cleanup hooks remain safe.

Closes: test-suite-integration-class-split
