+++ changelog.d/test-suite-integration-class-split.md
@@ -1,0 +1,5 @@
+---
+category: Maintenance
+---
+
+- **Maintenance:** split `IntegrationTests` kitchen-sink class into three focused classes (`IntegrationTests_WorkspaceCore`, `IntegrationTests_SymbolNavigation`, `IntegrationTests_EditApply`) — no assertions changed.
